+++ src/main/resources/views/TaskCell.fxml
@@ -6,8 +6,6 @@
 <?import javafx.scene.text.*?>
 <?import javafx.scene.web.*?>
 
-<!--<BorderPane prefHeight="150.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="views.TaskCell">-->
-
 <BorderPane prefHeight="150.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
    <center>
       <WebView disable="true" opacity="0.8" prefHeight="200.0" prefWidth="200.0" BorderPane.alignment="CENTER" />
